Rename NewGoto.srv to Goto.srv and add keep_position field in srv Goto.srv
For the captain to understand that a goto request must never finish
even if the accuracy has been accepted we use to use the tolerance:
    * If tolerance is 0.0 the waypoint never finalizes (timeout >= 3600s)

However, to avoid problems if someone forgets to add the tolerance
(by default is 0.0). we have add an extra field to clearly indicate
if the asked goto must keep the position or not.